Add delays

+++ clear_temp_folder.py
@@ -11,7 +11,7 @@ def clear_temp_folder():
     for element in tempfolder_content:
         element_path = os.path.join(tempfolder, element)
         for template in pathname_templates:
-            if os.path.isdir(element_path) and template in element:
+            if os.path.isdir(element_path) and template in element and 'snap-private-tmp' not in element:
                 shutil.rmtree(element_path)
     print('Temp folder cleared.\n')
